[exporter/datadogexporter] Move attribute functions to `attributes` package
