+++ exporter/datadogexporter/internal/attributes/azure/azure.go
