+++ exporter/datadogexporter/internal/metadata/ec2/ec2.go
@@ -14,20 +14,15 @@
 package ec2
 
 import (
-	"fmt"
 	"strings"
 
 	"github.com/aws/aws-sdk-go/aws/ec2metadata"
 	"github.com/aws/aws-sdk-go/aws/session"
-	"go.opentelemetry.io/collector/model/pdata"
-	conventions "go.opentelemetry.io/collector/model/semconv/v1.5.0"
 	"go.uber.org/zap"
 )
 
 var (
-	defaultPrefixes  = [3]string{"ip-", "domu", "ec2amaz-"}
-	ec2TagPrefix     = "ec2.tag."
-	clusterTagPrefix = ec2TagPrefix + "kubernetes.io/cluster/"
+	defaultPrefixes = [3]string{"ip-", "domu", "ec2amaz-"}
 )
 
 type HostInfo struct {
@@ -85,57 +80,4 @@ func (hi *HostInfo) GetHostname(logger *zap.Logger) string {
 	}
 
 	return hi.EC2Hostname
-}
-
-// HostnameFromAttributes gets a valid hostname from labels
-// if available
-func HostnameFromAttributes(attrs pdata.AttributeMap) (string, bool) {
-	hostName, ok := attrs.Get(conventions.AttributeHostName)
-	if ok && !isDefaultHostname(hostName.StringVal()) {
-		return hostName.StringVal(), true
-	}
-
-	if hostID, ok := attrs.Get(conventions.AttributeHostID); ok {
-		return hostID.StringVal(), true
-	}
-
-	return "", false
-}
-
-// HostInfoFromAttributes gets EC2 host info from attributes following
-// OpenTelemetry semantic conventions
-func HostInfoFromAttributes(attrs pdata.AttributeMap) (hostInfo *HostInfo) {
-	hostInfo = &HostInfo{}
-
-	if hostID, ok := attrs.Get(conventions.AttributeHostID); ok {
-		hostInfo.InstanceID = hostID.StringVal()
-	}
-
-	if hostName, ok := attrs.Get(conventions.AttributeHostName); ok {
-		hostInfo.EC2Hostname = hostName.StringVal()
-	}
-
-	attrs.Range(func(k string, v pdata.AttributeValue) bool {
-		if strings.HasPrefix(k, ec2TagPrefix) {
-			tag := fmt.Sprintf("%s:%s", strings.TrimPrefix(k, ec2TagPrefix), v.StringVal())
-			hostInfo.EC2Tags = append(hostInfo.EC2Tags, tag)
-		}
-		return true
-	})
-
-	return
-}
-
-// ClusterNameFromAttributes gets the AWS cluster name from attributes
-func ClusterNameFromAttributes(attrs pdata.AttributeMap) (clusterName string, ok bool) {
-	// Get cluster name from tag keys
-	// https://github.com/DataDog/datadog-agent/blob/1c94b11/pkg/util/ec2/ec2.go#L238
-	attrs.Range(func(k string, _ pdata.AttributeValue) bool {
-		if strings.HasPrefix(k, clusterTagPrefix) {
-			clusterName = strings.Split(k, "/")[2]
-			ok = true
-		}
-		return true
-	})
-	return
 }
