+++ exporter/datadogexporter/internal/metadata/host_test.go
@@ -20,12 +20,9 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	conventions "go.opentelemetry.io/collector/model/semconv/v1.5.0"
 	"go.uber.org/zap"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/datadogexporter/config"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/datadogexporter/internal/metadata/azure"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/datadogexporter/internal/testutils"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/exporter/datadogexporter/internal/utils/cache"
 )
 
@@ -61,151 +58,4 @@ func TestHost(t *testing.T) {
 	require.NoError(t, err)
 	// TODO: Investigate why the returned host contains more data on github actions.
 	assert.Contains(t, host, osHostname)
-}
-
-const (
-	testHostID      = "example-host-id"
-	testHostName    = "example-host-name"
-	testContainerID = "example-container-id"
-	testClusterName = "clusterName"
-	testNodeName    = "nodeName"
-	testCustomName  = "example-custom-host-name"
-)
-
-func TestHostnameFromAttributes(t *testing.T) {
-	// Custom hostname
-	attrs := testutils.NewAttributeMap(map[string]string{
-		AttributeDatadogHostname:            testCustomName,
-		AttributeK8sNodeName:                testNodeName,
-		conventions.AttributeK8SClusterName: testClusterName,
-		conventions.AttributeContainerID:    testContainerID,
-		conventions.AttributeHostID:         testHostID,
-		conventions.AttributeHostName:       testHostName,
-	})
-	hostname, ok := HostnameFromAttributes(attrs)
-	assert.True(t, ok)
-	assert.Equal(t, hostname, testCustomName)
-
-	// Container ID
-	attrs = testutils.NewAttributeMap(map[string]string{
-		conventions.AttributeContainerID: testContainerID,
-	})
-	hostname, ok = HostnameFromAttributes(attrs)
-	assert.True(t, ok)
-	assert.Equal(t, hostname, testContainerID)
-
-	// AWS cloud provider means relying on the EC2 function
-	attrs = testutils.NewAttributeMap(map[string]string{
-		conventions.AttributeCloudProvider: conventions.AttributeCloudProviderAWS,
-		conventions.AttributeHostID:        testHostID,
-		conventions.AttributeHostName:      testHostName,
-	})
-	hostname, ok = HostnameFromAttributes(attrs)
-	assert.True(t, ok)
-	assert.Equal(t, hostname, testHostName)
-
-	// GCP cloud provider means relying on the GCP function
-	attrs = testutils.NewAttributeMap(map[string]string{
-		conventions.AttributeCloudProvider: conventions.AttributeCloudProviderGCP,
-		conventions.AttributeHostID:        testHostID,
-		conventions.AttributeHostName:      testHostName,
-	})
-	hostname, ok = HostnameFromAttributes(attrs)
-	assert.True(t, ok)
-	assert.Equal(t, hostname, testHostName)
-
-	// Azure cloud provider means relying on the Azure function
-	attrs = testutils.NewAttributeMap(map[string]string{
-		conventions.AttributeCloudProvider: conventions.AttributeCloudProviderAzure,
-		conventions.AttributeHostID:        testHostID,
-		conventions.AttributeHostName:      testHostName,
-	})
-	hostname, ok = HostnameFromAttributes(attrs)
-	assert.True(t, ok)
-	assert.Equal(t, hostname, testHostName)
-
-	// Host Id takes preference
-	attrs = testutils.NewAttributeMap(map[string]string{
-		conventions.AttributeHostID:   testHostID,
-		conventions.AttributeHostName: testHostName,
-	})
-	hostname, ok = HostnameFromAttributes(attrs)
-	assert.True(t, ok)
-	assert.Equal(t, hostname, testHostID)
-
-	// No labels means no hostname
-	attrs = testutils.NewAttributeMap(map[string]string{})
-	hostname, ok = HostnameFromAttributes(attrs)
-	assert.False(t, ok)
-	assert.Empty(t, hostname)
-}
-
-func TestGetClusterName(t *testing.T) {
-	// OpenTelemetry convention
-	attrs := testutils.NewAttributeMap(map[string]string{
-		conventions.AttributeK8SClusterName: testClusterName,
-	})
-	cluster, ok := getClusterName(attrs)
-	assert.True(t, ok)
-	assert.Equal(t, cluster, testClusterName)
-
-	// Azure
-	attrs = testutils.NewAttributeMap(map[string]string{
-		conventions.AttributeCloudProvider: conventions.AttributeCloudProviderAzure,
-		azure.AttributeResourceGroupName:   "MC_aks-kenafeh_aks-kenafeh-eu_westeurope",
-	})
-	cluster, ok = getClusterName(attrs)
-	assert.True(t, ok)
-	assert.Equal(t, cluster, "aks-kenafeh-eu")
-
-	// AWS
-	attrs = testutils.NewAttributeMap(map[string]string{
-		conventions.AttributeCloudProvider:          conventions.AttributeCloudProviderAWS,
-		"ec2.tag.kubernetes.io/cluster/clustername": "dummy_value",
-	})
-	cluster, ok = getClusterName(attrs)
-	assert.True(t, ok)
-	assert.Equal(t, cluster, "clustername")
-
-	// None
-	attrs = testutils.NewAttributeMap(map[string]string{})
-	_, ok = getClusterName(attrs)
-	assert.False(t, ok)
-}
-
-func TestHostnameKubernetes(t *testing.T) {
-	// Node name and cluster name
-	attrs := testutils.NewAttributeMap(map[string]string{
-		AttributeK8sNodeName:                testNodeName,
-		conventions.AttributeK8SClusterName: testClusterName,
-		conventions.AttributeContainerID:    testContainerID,
-		conventions.AttributeHostID:         testHostID,
-		conventions.AttributeHostName:       testHostName,
-	})
-	hostname, ok := HostnameFromAttributes(attrs)
-	assert.True(t, ok)
-	assert.Equal(t, hostname, "nodeName-clusterName")
-
-	// Node name, no cluster name
-	attrs = testutils.NewAttributeMap(map[string]string{
-		AttributeK8sNodeName:             testNodeName,
-		conventions.AttributeContainerID: testContainerID,
-		conventions.AttributeHostID:      testHostID,
-		conventions.AttributeHostName:    testHostName,
-	})
-	hostname, ok = HostnameFromAttributes(attrs)
-	assert.True(t, ok)
-	assert.Equal(t, hostname, "nodeName")
-
-	// no node name, cluster name
-	attrs = testutils.NewAttributeMap(map[string]string{
-		conventions.AttributeK8SClusterName: testClusterName,
-		conventions.AttributeContainerID:    testContainerID,
-		conventions.AttributeHostID:         testHostID,
-		conventions.AttributeHostName:       testHostName,
-	})
-	hostname, ok = HostnameFromAttributes(attrs)
-	assert.True(t, ok)
-	// cluster name gets ignored, fallback to next option
-	assert.Equal(t, hostname, testHostID)
 }
